backend/requirements.txt
os
dotenv
mysql-connector-python
alpaca-trade-api
datetime
requests
asyncio
fastapi
uvicorn